Refactor: 불필요한 depth 제거 후 Chat에 로직 이전
