+++ src/components/common/Chat/index.tsx
@@ -1,19 +1,35 @@
-import { useRef } from "react";
-import { useAtomValue } from "jotai";
-import { getSessionStorage, setSessionStorage } from "@/utils/sessionStorage";
-import { cn } from "@/utils/tailwind";
-import ChatSection from "./ChatSection";
-import ChatPeek from "./ChatPeek";
-import { $isChatOpen } from "@/store/chat";
+import { useEffect, useRef } from "react";
+import { getSessionStorage } from "@/utils/sessionStorage";
+import MessagePeek from "./MessagePeek";
+import Message from "./Message";
+
+// TODO: [2024.02.23] 더미 데이터 제거 및 WS 연결
+const DUMMY_MESSAGES = [
+  {
+    src: "https://i.pinimg.com/564x/93/66/aa/9366aa85149388c2e1f09bbf9d79b60e.jpg",
+    isMyMessage: false,
+  },
+  {
+    src: "https://i.pinimg.com/564x/dc/77/a8/dc77a8b6e9dd0a4a17470a0e2f85cd9f.jpg",
+    isMyMessage: false,
+  },
+  {
+    src: "https://i.pinimg.com/564x/bb/26/c6/bb26c6670b60beff3d81ef74771f2c69.jpg",
+    isMyMessage: true,
+  },
+  {
+    src: "https://jjalbang.today/files/jjalbox/2019/01/20190120_5c435f3d5c0d4.jpg",
+    isMyMessage: false,
+  },
+  {
+    src: "https://jjalbang.today/files/jjalbox/2021/06/20210608_60bf89513e1ea.jpg",
+    isMyMessage: false,
+  },
+];
 
 const Chat = () => {
-  const isChatOpen = useAtomValue($isChatOpen);
   const chatRoomRef = useRef<HTMLDivElement>(null);
 
-  const handleScroll = () => {
-    if (!chatRoomRef.current) return;
-    setSessionStorage("sidebar-scroll", chatRoomRef.current?.scrollTop);
-  };
   const setScrollPosition = () => {
     if (!chatRoomRef.current) return;
     const top = getSessionStorage("sidebar-scroll");
@@ -22,16 +38,24 @@ const Chat = () => {
     }
   };
 
+  useEffect(() => {
+    setScrollPosition();
+  }, [setScrollPosition]);
+
   return (
-    <div
-      className={cn(
-        "absolute bottom-0 left-0 flex h-3/5 w-full flex-col transition-transform sm:static sm:h-full sm:w-auto sm:translate-y-0 sm:flex-row",
-        isChatOpen ? "-translate-y-40pxr" : "translate-y-[calc(100%-40px)]",
-      )}
-    >
-      <ChatPeek handleScroll={handleScroll} />
-      {isChatOpen && <ChatSection ref={chatRoomRef} setScrollPosition={setScrollPosition} />}
-    </div>
+    <section className="relative h-full w-full">
+      <div
+        ref={chatRoomRef}
+        className="relative h-full w-full overflow-y-auto rounded-16pxr bg-background pb-15pxr"
+      >
+        <div className="flex flex-1 flex-col ">
+          {DUMMY_MESSAGES.map(({ src, isMyMessage }, index) => (
+            <Message key={`${index}-${src}`} src={src} isMyMessage={isMyMessage} />
+          ))}
+        </div>
+      </div>
+      <MessagePeek />
+    </section>
   );
 };
